requirements/mmedge_install.txt
addict==2.4.0
backcall==0.2.0
certifi==2021.10.8
cycler==0.11.0
Cython==0.29.25
dataclasses==0.6
decorator==5.1.0
fonttools==4.28.3
future==0.18.2
ipdb==0.13.9
ipython==7.30.1
jedi==0.18.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mmcv-full==1.4.0
-e git+http://github.com/open-mmlab/mmdetection.git@f08548bfd6d394a82566022709b5ce9e6b0a855e#egg=mmdet
numpy==1.21.4
opencv-python==4.5.2.52
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycocotools==2.0.3
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
terminaltables==3.1.10
toml==0.10.2
torch==1.7.0+cu110
torchvision==0.8.1+cu110
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
yapf==0.31.0
